Help Me Hardcore 3.1 Literal Hell Update

+++ src/main/java/net/romeosnowblitz/hmh2/fluid/magic_fluids/AbsorptionFluid.java
@@ -41,7 +41,7 @@ public abstract class AbsorptionFluid extends FlowableFluid {
         }
 
         public int getLevel(FluidState state) {
-            return (Integer)state.get(LEVEL);
+            return state.get(LEVEL);
         }
 
         public boolean isStill(FluidState state) {
